+++ src/main/java/org/jukeboxmc/block/behavior/BlockBell.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.data.Attachment;
 import org.jukeboxmc.block.direction.BlockFace;
@@ -56,7 +56,7 @@ public class BlockBell extends Block {
     }
 
     public void setToggle( boolean value ) {
-        this.setState( "toggle_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "toggle_bit", value ? 1 : 0 );
     }
 
     public boolean isToggle() {
