+++ src/main/java/org/jukeboxmc/block/behavior/BlockCoral.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.data.CoralColor;
 import org.jukeboxmc.item.Item;
@@ -36,7 +36,7 @@ public class BlockCoral extends Block {
     }
 
     public BlockCoral setDead( boolean value ) {
-        return this.setState( "dead_bit", value ? (byte) 1 : (byte) 0 );
+        return this.setState( "dead_bit", value ? 1 : 0 );
     }
 
     public boolean isDead() {
