+++ src/main/java/org/jukeboxmc/block/behavior/BlockDropper.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.direction.BlockFace;
 import org.jukeboxmc.blockentity.BlockEntity;
@@ -51,7 +51,7 @@ public class BlockDropper extends Block {
     }
 
     public void setTriggered( boolean value ) {
-        this.setState( "triggered_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "triggered_bit", value ? 1 : 0 );
     }
 
     public boolean isTriggered() {
