+++ src/main/java/org/jukeboxmc/block/behavior/BlockEndPortalFrame.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.direction.BlockFace;
 import org.jukeboxmc.block.direction.Direction;
@@ -31,7 +31,7 @@ public class BlockEndPortalFrame extends Block {
     }
 
     public void setEndPortalEye( boolean value ) {
-        this.setState( "end_portal_eye_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "end_portal_eye_bit", value ? 1 : 0 );
     }
 
     public boolean isEndPortalEye() {
