+++ src/main/java/org/jukeboxmc/block/behavior/BlockHopper.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.direction.BlockFace;
 import org.jukeboxmc.blockentity.BlockEntity;
@@ -53,7 +53,7 @@ public class BlockHopper extends Block {
     }
 
     public void setToggle( boolean value ) {
-        this.setState( "toggle_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "toggle_bit", value ? 1 : 0 );
     }
 
     public boolean isToggle() {
