+++ src/main/java/org/jukeboxmc/block/behavior/BlockLantern.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.direction.BlockFace;
 import org.jukeboxmc.item.Item;
@@ -61,7 +61,7 @@ public class BlockLantern extends Block {
     }
 
     public void setHanging( boolean value ) {
-        this.setState( "hanging", value ? (byte) 1 : (byte) 0 );
+        this.setState( "hanging", value ? 1 : 0 );
     }
 
     public boolean isHanging() {
