+++ src/main/java/org/jukeboxmc/block/behavior/BlockSapling.java
@@ -1,6 +1,6 @@
 package org.jukeboxmc.block.behavior;
 
-import com.nukkitx.nbt.NbtMap;
+import org.cloudburstmc.nbt.NbtMap;
 import org.jukeboxmc.block.Block;
 import org.jukeboxmc.block.data.SaplingType;
 import org.jukeboxmc.item.Item;
@@ -36,7 +36,7 @@ public class BlockSapling extends Block {
     }
 
     public void setAge( boolean value ) {
-        this.setState( "age_bit", value ? (byte) 1 : (byte) 0 );
+        this.setState( "age_bit", value ? 1 : 0 );
     }
 
     public boolean hasAge() {
